eip2718test: add eip 2718 parlia test
